refactor(layouts): rename 'model' folder to 'config'
