+++ src/widgets/layout/config/constants.ts
