+++ src/widgets/layout/config/index.ts
